🔨 Create pre-push hook and reduce validation in pre-commit
